Incomplete implementation for #15
Missing notifications for "opponent accepted your challenge" and
"opponent resigned". I've tried putting the acceptance notification
everywhere I could think to, but the notification never fires. I know
where the resgination notif. should go and that's labelled with an `XX:
notification:`, but I can't think why it wouldn't fire.

Other than their placement in the file, the only notable thing about
these notifications is the `path` in `hark-store` they're sent to. Most
notifications are sent to `/challenges`, `/games/<game-id>/moves`, or
`/games/<game-id>/updates`, but to avoid referring to games that don't
exist the acceptance notif. is sent to `/challenges` and the resignation
is sent to `/games`. I don't think this should matter as those locations
are created by referencing them in the `hark-store` poke, so maybe I'm
trying to fire them in the wrong place.

I'm aware `hark-store` is being deprecated, so I've left comments
sketching more granular notifications for moves, checks, and checkmates, but I think
that'll require some work across the codebase that's not worth doing if
`hark-store` is going to be out of date in a couple of months. Maybe
there should be a 'checkmate!' notification, but as mentioned in the
comments around line 900 how we solve issue #3 might interfere with that
logic.