Move styling from TSX to CSS, add responsiveness
The exception to this is the text above the chessboard, which is
currently hard-coded in TSX depending on the length of the @p. That
information will likely be moved to the Game Panel over the course of
the bounty.

This commit also adds responsive.css, and makes some minimal changes
to support a viewport width as low as 1080px, below which the UI
starts to break with its current styling.