Add skeleton of new frontend
As we start to merge backend features, we need to be able to plug them
into something resembling the Chess v1.0 frontend.

This commit partially implements the sketches we've discussed so far,
just enough to keep us moving. None of the added UI is functional,
added functionality will change the way it looks.

The focus is on new buttons/menus corresponding to what we're working
on over the next couple of weeks. Some minor visual tweaks have been
made (e.g. capitalizing some UI text makes it more legible) but
nothing here is remotely final.