Add panel to view previous positions
To feel something like a feature-complete Chess app, %chess needs an
interface for viewing previous positions in a game.

This commit refactors position updates to include Standard Algebraic
Notation and Forsyth-Edwards Notation on the move/new position they
represent. This information is stored in the frontend and the left-hand
"game panel" uses it to display previous moves and positions.