List outgoing challenges on the frontend (#45)
This commit fills the "Outgoing" list on the frontend with our outgoing
challenges. In doing so, it refactors the way challenges are sent and
received.

All challenge updates travel on one /challenges wire, and
there are now four challenge cards which are only sent to our frontend
if challenges/responses have been acked by our opponent. This prevents
frontend and backend state from diverging, which ensures the UI is
correctly informing the player about what's going on.